requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
pysaml2
requests
